+++ CHANGELOG.md
@@ -8,3 +8,4 @@ __Breaking Changes:__
 __New Features and Enhancements:__
 
 - Add support for the uploader display name field for Files and File Versions ([#424](https://github.com/box/box-android-sdk/pull/424))
+- Add path parameter sanitization ([#428](https://github.com/box/box-android-sdk/pull/428))